+++ pycricinfo/api/endpoints/wrapper.py
@@ -2,8 +2,16 @@ import requests
 from fastapi import APIRouter, Depends, Path, Query, status
 
 from pycricinfo.output_models.scorecard import CricinfoScorecard
-from pycricinfo.search.call_cricinfo_api import get_match_basic, get_play_by_play, get_player, get_scorecard, get_team
+from pycricinfo.search.call_cricinfo_api import (
+    get_match,
+    get_match_basic,
+    get_play_by_play,
+    get_player,
+    get_scorecard,
+    get_team,
+)
 from pycricinfo.source_models.api.commentary import CommentaryItem
+from pycricinfo.source_models.api.match import Match
 from pycricinfo.source_models.api.match_basic import MatchBasic
 from pycricinfo.source_models.api.player import Player
 from pycricinfo.source_models.api.team import TeamFull
@@ -57,12 +65,25 @@ async def get_match_team(
 
 
 @router.get(
-    "/match_summary/{match_id}",
+    "/match_summary/{series_id}/{match_id}",
     responses={status.HTTP_200_OK: {"description": "The match summary"}},
     summary="Get a match summary",
 )
-async def scorecard(match_id: int = Path(description="The Match ID")) -> CricinfoScorecard:
-    return get_scorecard(match_id)
+async def match(
+    series_id: int = Path(description="The Series ID"), match_id: int = Path(description="The Match ID")
+) -> Match:
+    return get_match(series_id, match_id)
+
+
+@router.get(
+    "/scorecard/{series_id}/{match_id}",
+    responses={status.HTTP_200_OK: {"description": "The match summary"}},
+    summary="Get a match summary",
+)
+async def scorecard(
+    series_id: int = Path(description="The Series ID"), match_id: int = Path(description="The Match ID")
+) -> CricinfoScorecard:
+    return get_scorecard(series_id, match_id)
 
 
 @router.get(