+++ pycricinfo/config.py
@@ -17,7 +17,7 @@ class CoreAPIRoutes(BaseModel):
     player: str = "teams/0/athletes/{player_id}"
     match_basic: str = "events/{match_id}"
     match_team: str = "leagues/0/events/{match_id}/competitions/{match_id}/competitors/{team_id}"
-    match_summary: str = "0/summary?event={match_id}"
+    match_summary: str = "{series_id}/summary?event={match_id}"
     league: str = "leagues/{league_id}"
     play_by_play_page: str = "0/playbyplay?event={match_id}&page={page}&period={innings}"
     venue: str = "venues/{venue_id}"